Add files via upload
I still think I didn't understand finding the y value with the dplyr functions. Are we supposed to be able to find the 0.5 value from the mutated column? Or am I overthinking it? I did not download the work that we did outside of class so I regret not being able to look at that again.